Fix bloated links, split into two link files. Update IDs
